refactor: extract workspace git operations
